+++ NEWS.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 - Several other minor bugs were resolved (#942, #943).
+- Type checking was not performed correctly for conversion function
+  actuals in port maps (#945).
 
 ## Version 1.13.2 - 2024-08-11
 - Fixed an incorrect bounds check error when a constant declaration has